move logic to auth slice
