+++ src/redux/userThunkSlice.tsx
@@ -1,9 +1,4 @@
-import {
-  ILoginRequest,
-  ISignupRequest,
-  IGuestSignupRequest,
-  IUserState,
-} from 'types/UserData.type';
+import { IUserState } from 'types/UserData.type';
 
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import axios from 'axios';
@@ -34,55 +29,6 @@ const initialState: IUserState = {
   },
   status: 'pending',
 };
-
-//create async payload callback function
-//LOGIN
-export const login = createAsyncThunk(
-  'login',
-  async (loginData: ILoginRequest, { rejectWithValue }) => {
-    try {
-      const result = await axios.post(
-        `${API_URL}/login`,
-        JSON.stringify(loginData)
-      );
-      return result.data;
-    } catch (error: any) {
-      return rejectWithValue(error.response.data);
-    }
-  }
-);
-
-//SIGNUP
-export const signup = createAsyncThunk(
-  'signup',
-  async (signupData: ISignupRequest, thunkAPI) => {
-    try {
-      const result = await axios.post(
-        `${API_URL}/signup`,
-        JSON.stringify(signupData)
-      );
-      return result.data;
-    } catch (error: any) {
-      return thunkAPI.rejectWithValue(error.response);
-    }
-  }
-);
-
-//GUEST SIGNUP
-export const signupGuest = createAsyncThunk(
-  'signup',
-  async (signupData: IGuestSignupRequest, thunkAPI) => {
-    try {
-      const result = await axios.post(
-        `${API_URL}/signup`,
-        JSON.stringify(signupData)
-      );
-      return result.data;
-    } catch (error: any) {
-      return thunkAPI.rejectWithValue(error.response);
-    }
-  }
-);
 
 //GET
 export const getUser = createAsyncThunk(
@@ -151,12 +97,6 @@ export const userSlice = createSlice({
   },
   extraReducers: (builder) => {
     builder
-      .addCase(login.fulfilled, (state, action) => {
-        state.status = 'pending';
-      })
-      .addCase(signup.fulfilled, (state, action) => {
-        state.status = 'pending';
-      })
       .addCase(getUser.fulfilled, (state, action) => {
         state.status = 'pending';
         state.user = action.payload;
