+++ src/views/components/organisms/GuestLoginForm.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useAppDispatch } from 'hooks/hooks';
 import { useNavigate } from 'react-router';
 import useForm from 'hooks/useForm';
-import { login } from 'redux/userThunkSlice';
+import { login } from 'redux/authSlice';
 import { SessionKeys } from 'constants/sessionKeys';
 import SessionServices from 'services/session.services';
 import { Form } from '../atoms/Form';
